% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_biofire.R
\name{clean_biofire_pdf}
\alias{clean_biofire_pdf}
\title{Read a Biofire GI PDF file to a dataframe}
\usage{
clean_biofire_pdf(file)
}
\arguments{
\item{infile}{Path to the input PDF file}
}
\value{
A dataframe of the results
}
\description{
This function loads a Biofire GI PDF output file.
It opens it with pdftools, uses regex to extract the results
and formats the output as a dataframe
}
