% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_biofire.R
\name{clean_biofire_txt}
\alias{clean_biofire_txt}
\title{Read a Biofire GI text file to a dataframe}
\usage{
clean_biofire_txt(file)
}
\arguments{
\item{infile}{Path to the input text file}
}
\value{
A dataframe of the results
}
\description{
This function loads a Biofire GI text output file.
It opens it with pdftools, uses regex to extract the results
and formats the output as a dataframe
}
